requirements.txt
# python >= 3.12.7
python-dotenv==1.0.1
openai==1.59.7
flask==3.1.0

